+++ src/BDSBendBuilder.cc
@@ -100,7 +100,19 @@ BDSAcceleratorComponent* BDS::BuildSBendLine(const G4String&         elementName
 	{buildFringeIncoming = true;}
       if (BDS::IsFinite(fintx) && BDS::IsFinite(hgap))
 	{buildFringeOutgoing = true;}      
-      
+
+    // don't build fringe elements if a non-matrix integrator set is used and
+    // only a pole face effect is specified. The angled pole face geometry is
+    // constructed so a thin integrated pole face kick shouldn't be applied as well
+    if (!integratorSet->IsMatrixIntegratorSet() && BDSGlobalConstants::Instance()->BuildPoleFaceGeometry())
+      {
+        // no hgap check - finite hgap but 0 fint/x will not produce a fringe kick so fringes won't be built anyway
+        if (BDS::IsFinite(element->e1) && !BDS::IsFinite(fint))
+          {buildFringeIncoming = false;}
+        if (BDS::IsFinite(element->e2) && !BDS::IsFinite(fintx))
+          {buildFringeOutgoing = false;}
+      }
+
       // overriding checks - don't build fringe field if we're butt against another
       // sbend.
       if (prevElement)
@@ -303,8 +315,16 @@ BDSAcceleratorComponent* BDS::BuildSBendLine(const G4String&         elementName
   // build incoming fringe field if required
   if (buildFringeIncoming)
     {
+      G4double e1 = element->e1;
+      // if using a non-matrix integrator set, this code should only be reached if a finite fint is specified.
+      // As the angled face geometry is constructed, a thin pole face kick shouldn't also be applied, so set
+      // e1 to 0 in the magnet strength object so only fringe effects are applied
+      if (!integratorSet->IsMatrixIntegratorSet() && BDSGlobalConstants::Instance()->BuildPoleFaceGeometry()
+            && BDS::IsFinite(fint))
+        {e1 = 0;}
+
       BDSMagnetStrength* fringeStIn = BDS::GetFringeMagnetStrength(element, st, oneFringeAngle,
-                                                                   element->e1, element->e2, fintx, true);
+                                                                   e1, element->e2, fintx, true);
       G4String segmentName           = baseName + "_e1_fringe";
       G4double fringeAngleIn         = 0.5*oneFringeAngle - incomingFaceAngle;
       G4double fringeAngleOut        = 0.5*oneFringeAngle + incomingFaceAngle;
@@ -402,8 +422,16 @@ BDSAcceleratorComponent* BDS::BuildSBendLine(const G4String&         elementName
   //Last element should be fringe if poleface specified
   if (buildFringeOutgoing)
     {
+      G4double e2 = element->e2;
+      // if using a non-matrix integrator set, this code should only be reached if a finite fintx is specified.
+      // As the angled face geometry is constructed, a thin pole face kick shouldn't also be applied, so set
+      // e2 to 0 in the magnet strength object so only fringe effects are applied
+      if (!integratorSet->IsMatrixIntegratorSet() && BDSGlobalConstants::Instance()->BuildPoleFaceGeometry()
+          && BDS::IsFinite(fintx))
+        {e2 = 0;}
+
       BDSMagnetStrength* fringeStOut = BDS::GetFringeMagnetStrength(element, st, oneFringeAngle,
-                                                                    element->e1, element->e2, fintx, false);
+                                                                    element->e1, e2, fintx, false);
       G4double fringeAngleIn          = 0.5*oneFringeAngle + outgoingFaceAngle;
       G4double fringeAngleOut         = 0.5*oneFringeAngle - outgoingFaceAngle;
       G4String segmentName            = baseName + "_e2_fringe";
@@ -668,8 +696,16 @@ BDSLine* BDS::BuildRBendLine(const G4String&         elementName,
   
   if (buildFringeIncoming)
     {
-      BDSMagnetStrength* fringeStIn = BDS::GetFringeMagnetStrength(element, st, oneFringeAngle,
-                                                                   trackingPolefaceAngleIn, trackingPolefaceAngleOut,
+      G4double trackingPolefaceAngle = trackingPolefaceAngleIn;
+      // if using a non-matrix integrator set, this code should only be reached if a finite fint is specified.
+      // As the angled face geometry is constructed, a thin pole face kick shouldn't also be applied, so subtract
+      // e1 for in the magnet strength object so only the fringe & the natural rbend angled face effects are applied
+      if (!integratorSet->IsMatrixIntegratorSet() && BDSGlobalConstants::Instance()->BuildPoleFaceGeometry()
+          && BDS::IsFinite(fint))
+        {trackingPolefaceAngle -= element->e1;}
+
+        BDSMagnetStrength* fringeStIn = BDS::GetFringeMagnetStrength(element, st, oneFringeAngle,
+                                                                     trackingPolefaceAngle, trackingPolefaceAngleOut,
                                                                    fintx, true);
       G4String fringeName = name + "_e1_fringe";
 
@@ -726,8 +762,15 @@ BDSLine* BDS::BuildRBendLine(const G4String&         elementName,
   //Last element should be fringe if poleface specified
   if (buildFringeOutgoing)
     {
+      G4double trackingPolefaceAngle = trackingPolefaceAngleOut;
+      // if using a non-matrix integrator set, this code should only be reached if a finite fintx is specified.
+      // As the angled face geometry is constructed, a thin pole face kick shouldn't also be applied, so subtract
+      // e2 for in the magnet strength object so only the fringe & the natural rbend angled face effects are applied
+      if (!integratorSet->IsMatrixIntegratorSet() && BDSGlobalConstants::Instance()->BuildPoleFaceGeometry()
+          && BDS::IsFinite(fintx))
+        {trackingPolefaceAngle -= element->e2;}
       BDSMagnetStrength* fringeStOut = BDS::GetFringeMagnetStrength(element, st, oneFringeAngle,
-                                                                    trackingPolefaceAngleIn, trackingPolefaceAngleOut,
+                                                                    trackingPolefaceAngleIn, trackingPolefaceAngle,
                                                                     fintx, false);
       G4String fringeName = name + "_e2_fringe";